Мои изменения

+++ src/main/java/Main.java
@@ -6,9 +6,9 @@ public class Main {
         Formatter formatter = new Formatter();
         Calculator calculator = new Calculator();
         calculator.calculate();
-        String str = "Добавленные товары:\n%sСумма с человека: %d %s";
+        String str = "Добавленные товары:\n%sСумма с человека: %.2f %s";
         String formatRubl = calculator.sumNameProduct;
-        int formatDouble = (int) (calculator.sumPriceProduct / persons.newpersone);
+        double formatDouble = calculator.sumPriceProduct / persons.newpersone;
         String s = formatter.padejRub(formatDouble);
         System.out.println(String.format(str, formatRubl, formatDouble, s));
     }
